Updated documentation about Django Nova
