+++ tests/conftest.py
@@ -1,23 +1,12 @@
-import django
-from django.conf import settings
+import pytest
 
-if not settings.configured:
-    settings.configure(
-        DEBUG=True,
-        SECRET_KEY="test-secret-key-nova-2025",
-        DATABASES={
-            "default": {
-                "ENGINE": "django.db.backends.sqlite3",
-                "NAME": ":memory:",
-            }
-        },
-        INSTALLED_APPS=[
-            "django.contrib.auth",
-            "django.contrib.contenttypes",
-            "nova",
-            "tests.apps.TestsConfig",
-        ],
-        USE_TZ=True,
-    )
-
-    django.setup()
+@pytest.fixture(scope="session")
+def django_db_setup(django_db_setup, django_db_blocker):
+    """
+    Redefining the creation of a database for the library.
+    Since we don't have a migrations folder in the tests/ folder, we use
+    the --run-syncdb flag, which creates tables directly from the model classes.
+    """
+    with django_db_blocker.unblock():
+        from django.core.management import call_command
+        call_command('migrate', '--run-syncdb', verbosity=0, interactive=False)
